+++ backend/src/main/java/com/ktnu/AiLectureSummary/config/SecurityConfig.java
@@ -33,8 +33,8 @@ public class SecurityConfig {
                 .csrf(csrf -> csrf.disable())  // API 서버는 보통 CSRF 비활성화
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)) // httponly 설정
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/api/members/login", "/api/members/register", "/health",
-                                "/v3/api-docs/**","/api/members/verify","/api/members/reset-password",
+                        .requestMatchers("/api/auth/login", "/api/auth/register", "/health",
+                                "/v3/api-docs/**","/api/password/verify","/api/password/reset",
                                 "/swagger-ui/**",
                                 "/swagger-ui.html").permitAll() // 로그인, 회원가입, 스웨거, 헬스체크, 비밀번호 변경 인증없이 접근 허용
 //                        .requestMatchers(HttpMethod.GET, "/api/Lecture/**").permitAll()
