+++ backend/src/test/java/com/ktnu/AiLectureSummary/service/MemberServiceTest.java
@@ -1,8 +1,19 @@
 package com.ktnu.AiLectureSummary.service;
 
+import com.ktnu.AiLectureSummary.repository.MemberRepository;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
 
 import static org.junit.jupiter.api.Assertions.*;
-class MemberServiceTest {
+import static org.mockito.Mockito.mock;
 
+class MemberServiceTest {
+    private MemberService memberService;
+    private MemberRepository memberRepository = mock(MemberRepository.class);
+
+    @BeforeEach
+    void setUp() {
+        memberService = new MemberService(memberRepository);
+
+    }
 }
