+++ docker-compose.prod.yml
@@ -55,6 +55,13 @@ services:
     # restart: always
     env_file:
       - ./.env.fastapi
+  redis:
+    image: redis:latest
+    container_name: redis
+    ports:
+      - "6379:6379"
+    # restart: always
+
 
 volumes:
   mysql-prod-data:
